+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Issue when running `sedge` without sudo.
+- Issue when `sedge` generates clients datadirs without sudo.
 
 ## [v1.2.0] - 2023-06-06